+++ kovshikov.nikita/F0/createGraph.hpp
@@ -1,18 +1,17 @@
 #ifndef CREATEGRAPH_HPP
 #define CREATEGRAPH_HP
 #include "orientedGraph.hpp"
+#include "workingGraph.hpp"
 
 namespace kovshikov
 {
-  bool isDigit(char ch);
 // для создания графа надо устроить проверку на уникальность ключа
   void createGraph(std::map< std::string, Graph >& graphsList, std::istream& is);
   void createLonely(std::map< std::string, Graph >& graphsList, std::istream& is);
   void deleteGraph(std::map< std::string, Graph >& graphsList, std::istream& is);
   void workWith(std::map< std::string, Graph >& graphsList, std::istream& is);
 
-//в отдельную ЕТ:
-  void add(Graph& graph, std::istream& is);
+  void outputGraphs(const std::map< std::string, Graph >& graphsList, std::ostream& out);
 }
 
 #endif
